requirements.txt
aiosqlite==0.19.0
discord.py==2.3.2
py-cord==2.4.1
tensorflow==2.15.0
pillow==10.1.0
flask==3.0.0